MAINT: Move file saving out of core module.
